+++ src/app/(top)/list/page.jsx
@@ -3,10 +3,10 @@ import { Title } from "../_components/title";
 export default function List() {
   return (
     <div>
-      <Title text="お買い物リスト" />
-      <p className="h-96 w-96 text-red-500">お買い物リスト</p>
-      <p className="h-96 w-96 text-red-500">お買い物リスト</p>
-      <p className="h-96 w-96 text-red-500">お買い物リスト</p>
+      <Title text="買い物リスト" />
+      <p className="h-96 w-96 text-red-500">買い物リスト</p>
+      <p className="h-96 w-96 text-red-500">買い物リスト</p>
+      <p className="h-96 w-96 text-red-500">買い物リスト</p>
     </div>
   );
 }
